backend/requirements.txt
fastapi[standard]
motor
pymongo
bcrypt
pyjwt
python-multipart
email-validator
python-dotenv
sendgrid
httpx
aiohttp
groq
